// lib/types/products.ts
export interface IProductDetails {
  processor?: string;
  screen?: string;
  operating_system?: string;
  ram?: string;
  ssd?: string;
  ports?: string;
  graphic?: string;
  warranty?: string;
  back_camera?: string;
  front_camera?: string;
  battery?: string;
  frequency_response?: string;
  microphone?: boolean;
  wireless?: boolean;
  wireless_standby_time?: boolean;
  connectionType?: string[];
  connectors?: string[];
  bluetooth?: boolean;
  noise_cancelling?: boolean;
  sound_isolating?: boolean;
}

export type TSlug = {
  _type: string;
  current: string;
};

export type TImage = {
  _key: string;
  _type: "image";
  asset: {
    _ref: string;
    _type: "reference";
  };
};

export interface IProduct {
  id: string; // Adicionado para garantir acesso ao ID do produto
  image: any;
  name: string;
  slug: TSlug;
  price: number;
  // Fields that may be returned from server-side props should be serializable (no undefined)
  discount: number | null;
  details: IProductDetails[];
  brand: string;
  category: string[];
  subCategory: string | null; // Adicionado para refletir subcategoria
  isOffer: boolean;
  registerDate: string | null;
  timeStamp: number | null;
  starRating: number;
  description: string | null; // Adicionado para exibir descrição do banco
}